saymore/2 experiment
Trying to run multiple jobs in a single process trying to discover if
the execute sequentially and if the order is changed. Seems like the
order is maintained in some of my experiments, however; it really sucks
since I'm doing all the waiting as well. I was hoping that mapping all
transmissions would send them into the ether after which I could pick
everything up quickly as soon as I'm ready.